Added Arduino Uno fastled 18 led test
A very simple piece of test code to check that led strips work. Use led strips and the fastled library.

Connect CI and DI to pin 2 and 3 (can't remember which one goes to which). Power to 5V and ground to GND. No need for a breadboard, just put them directly on an Arduino Uno.

Remember to set the total number of pins.